added error handling in scripts

+++ deploy/scripts/install_fonts.sh
@@ -8,5 +8,11 @@ wget -O roboto.zip "https://fonts.google.com/download?family=Roboto"
 
 sudo apt install unzip -y
 
-unzip roboto.zip -d .fonts/
-rm roboto.zip
+if [ 1 -eq $( ls | grep -c "roboto.zip" ) ]
+then
+    unzip roboto.zip -d .fonts/
+    rm roboto.zip
+else
+    echo "Fonts download failed"
+    exit 1
+fi
